requirements.txt
alembic==1.0.11
aniso8601==7.0.0
asn1crypto==0.24.0
blinker==1.4
cairocffi===file-.cairocffi-VERSION
CairoSVG==2.4.1
cffi==1.12.3
Click==7.0
coverage==4.5.4
cryptography==2.7
cssselect2==0.2.1
defusedxml==0.6.0
dominate==2.4.0
first==2.0.2
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.7.2
Flask-Cors==3.0.8
Flask-GraphQL==2.0.0
Flask-JWT-Extended==3.21.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
Flask-WeasyPrint==0.6
Flask-WTF==0.14.2
graphene==2.1.8
graphene-file-upload==1.2.2
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
html5lib==1.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
olefile==0.46
pdfkit==0.6.1
pdfrw==0.4
piexif==1.1.3
Pillow==6.1.0
pip-tools==4.0.0
promise==2.2.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
Pyphen==0.9.5
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.7
tinycss2==1.0.2
typing==3.7.4.1
visitor==0.1.3
WeasyPrint==48
webencodings==0.5.1
Werkzeug==0.15.5
WTForms==2.2.1
